requirements.txt
Django==2.2
django-tinymce==3.0.2
django-embed-video==1.3.1